Refactor: move ResultWithException
