+++ src/Bicep.Core.UnitTests/Registry/CachedModules.cs
@@ -7,6 +7,7 @@ using System.Text.Json.Nodes;
 using System.Text.Json.Serialization;
 using Bicep.Core.Diagnostics;
 using Bicep.Core.SourceCode;
+using Bicep.Core.Utils;
 using FluentAssertions;
 
 namespace Bicep.Core.UnitTests.Registry;
