+++ src/Bicep.LangServer.IntegrationTests/CodeLensTests.cs
@@ -5,7 +5,6 @@ using System.Collections.Immutable;
 using System.Diagnostics.CodeAnalysis;
 using System.Reactive.Linq;
 using System.Reflection;
-using Bicep.Core.Diagnostics;
 using Bicep.Core.Registry;
 using Bicep.Core.Samples;
 using Bicep.Core.SourceCode;
@@ -13,6 +12,7 @@ using Bicep.Core.UnitTests;
 using Bicep.Core.UnitTests.Assertions;
 using Bicep.Core.UnitTests.Mock;
 using Bicep.Core.UnitTests.Utils;
+using Bicep.Core.Utils;
 using Bicep.LangServer.IntegrationTests.Helpers;
 using Bicep.LanguageServer.Handlers;
 using FluentAssertions;
